scaladoc: Error, ErrorMessage, Path, PathElement
