+++ chimney/src/main/scala/io/scalaland/chimney/partial/PathElement.scala
@@ -1,27 +1,67 @@
 package io.scalaland.chimney.partial
 
+/** Data type for representing path element in a (possibly) nested object structure.
+  *
+  * @since 0.7.0
+  */
 sealed trait PathElement {
+
+  /** Obtain conventional string representation of path element\
+    *
+    * @since 0.7.0
+    */
   def asString: String
 }
 
 object PathElement {
+
+  /** Object property accessor (e.g case class param name).
+    *
+    * @param name field name
+    *
+    * @since 0.7.0
+    */
   final case class Accessor(name: String) extends PathElement {
     override def asString: String = name
   }
 
+  /** Index in a collection.
+    *
+    * @param index index of an element
+    *
+    * @since 0.7.0
+    */
   final case class Index(index: Int) extends PathElement {
     override def asString: String = s"($index)"
   }
 
+  /** Value in the map.
+    *
+    * @param key key in the map
+    *
+    * @since 0.7.0
+    */
   final case class MapValue(key: Any) extends PathElement {
     override def asString: String = s"($key)"
   }
 
+  /** Key in the map.
+    *
+    * @param key key in the map
+    *
+    * @since 0.7.0
+    */
   final case class MapKey(key: Any) extends PathElement {
     override def asString: String = s"keys($key)"
   }
 
-  final def shouldPrependWithDot(pe: PathElement): Boolean = pe match {
+  /** Specifies if path element in conventional string representation should be prepended with a dot.
+    *
+    * @param pathElement path element
+    *
+    * @since 0.7.0
+    */
+  final def shouldPrependWithDot(pathElement: PathElement): Boolean = pathElement match {
     case _: Accessor => true
     case _: Index    => false
     case _: MapValue => false
